html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-pointer-move-source-into-pointer.html

<!doctype html>
<title>pointer updates (moveBefore source into the pointer)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<div id="holder"><source id="source-x"></div>
<script>
var events = [];
var a = 0;
var b = 0;
var c = 0;
var x = 0;
</script>
<video id="v"><source id="source-a" onerror="a++; events.push('a')"><source id="source-b" onerror="b++; events.push('b')" src="resources/delayed-broken-video.py"><source id="source-c" onerror="c++; events.push('c')"></video>
<script>
async_test(function(t) {
  var video = document.getElementById('v');
  assert_implements(typeof video.moveBefore == 'function', 'ParentNode.moveBefore() is required for this test');

  var sourceX = document.getElementById('source-x');
  sourceX.onerror = function() { x++; events.push('x'); };

  // Move an existing source from another parent to the pointer position: b | c.
  video.moveBefore(sourceX, document.getElementById('source-c'));

  window.onload = t.step_func(function() {
    assert_equals(a, 1, 'error events on a');
    assert_equals(b, 1, 'error events on b');
    assert_equals(x, 1, 'error events on x');
    assert_equals(c, 1, 'error events on c');
    assert_array_equals(events, ['a', 'b', 'x', 'c'], 'processed source order');
    t.done();
  });
});
</script>
